demo/src/main/java/junit/TriangleClassifier.java
package junit;

public class TriangleClassifier {
    public static String classifyTriangle(int side1, int side2, int side3) {
        if (!isValidTriangle(side1, side2, side3)) {
            return "NotATriangle";
        } else if (side1 == side2 && side2 == side3) {
            return "Equilateral";
        } else if (side1 == side2 || side1 == side3 || side2 == side3) {
            return "Isosceles";
        } else {
            return "Scalene";
        }
    }
    
    private static boolean isValidTriangle(int side1, int side2, int side3) {
        return (side1 + side2 > side3) && (side1 + side3 > side2) && (side2 + side3 > side1);
    }
}
